#1 [refactor] mapper, useCase 위치 이동 및 네이밍 변경
