--- core/src/main/java/com/sopetit/core/BaseMapper.kt
@@ -1,5 +1,0 @@
-package com.sopetit.core
-
-interface BaseMapper<RESPONSE, MODEL> {
-    fun mapToModel(response: RESPONSE?): MODEL
-}
