+++ core/src/main/java/com/sopetit/core/base/Mapper.kt
@@ -1,0 +1,5 @@
+package com.sopetit.core.base
+
+interface Mapper<RESPONSE, MODEL> {
+    fun mapToModel(response: RESPONSE?): MODEL
+}
