+++ core/src/main/java/com/sopetit/core/base/UseCase.kt
@@ -1,7 +1,7 @@
-package com.sopetit.core
+package com.sopetit.core.base
 
 import kotlinx.coroutines.flow.Flow
 
-abstract class BaseUseCase<REQUEST, RESPONSE> {
+abstract class UseCase<REQUEST, RESPONSE> {
     abstract suspend operator fun invoke(request: REQUEST): Flow<RESPONSE>
 }
